Arreglado provisionalmente la búsqueda de usuarios.
Digo provisionalmente ya que no salta error al poner un usuario erróneo, simplemente te redirige a la vista de búsqueda sin ningún mensaje. Puesto así por ahora por muchos problemas que está causando la validación.
